Migrate encryption perf cases to ote
